refactor: extract managed-process utilities into shared module
