+++ scripts/services/fpc-services-smoke.ts
@@ -1,4 +1,3 @@
-import { type ChildProcessWithoutNullStreams, spawn } from "node:child_process";
 import { mkdtempSync, readFileSync, rmSync, writeFileSync } from "node:fs";
 import os from "node:os";
 import path from "node:path";
@@ -14,7 +13,7 @@ import {
 import { Contract } from "@aztec/aztec.js/contracts";
 import { Fr } from "@aztec/aztec.js/fields";
 import { waitForL1ToL2MessageReady } from "@aztec/aztec.js/messaging";
-import { createAztecNodeClient, waitForNode } from "@aztec/aztec.js/node";
+import { createAztecNodeClient } from "@aztec/aztec.js/node";
 import {
   FeeJuiceContract,
   ProtocolContractAddress,
@@ -40,6 +39,17 @@ import {
 } from "viem";
 import { privateKeyToAccount } from "viem/accounts";
 
+import {
+  installManagedProcessSignalHandlers,
+  type ManagedProcess,
+  sleep,
+  startManagedProcess,
+  stopManagedProcess,
+  waitForHealth,
+  waitForLog,
+  waitForNodeReady,
+} from "../common/managed-process.ts";
+
 const DEFAULT_LOCAL_L1_PRIVATE_KEY =
   "0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80";
 const HEX_32_BYTE_PATTERN = /^0x[0-9a-fA-F]{64}$/;
@@ -54,11 +64,6 @@ const FEE_JUICE_PORTAL_ABI = parseAbi([
 ]);
 const QUOTE_DOMAIN_SEPARATOR = Fr.fromHexString("0x465043");
 
-type ManagedProcess = {
-  name: string;
-  process: ChildProcessWithoutNullStreams;
-  getLogs: () => string;
-};
 type TopupBridgeOutcome = "confirmed" | "timeout" | "failed";
 type TopupBridgeSubmission = {
   messageHash: Hex;
@@ -117,9 +122,6 @@ type ServiceFlowResult = {
   quoteSigBytes: number[];
 };
 
-const managedProcessRegistry = new Set<ManagedProcess>();
-let shutdownInProgress = false;
-
 function parseSmokeMode(value: string | undefined): SmokeMode {
   const normalized = (value ?? "both").trim().toLowerCase();
   if (
@@ -162,10 +164,6 @@ function readOptionalEnvBigInt(name: string): bigint | null {
 
 function ceilDiv(numerator: bigint, denominator: bigint): bigint {
   return (numerator + denominator - 1n) / denominator;
-}
-
-function sleep(ms: number): Promise<void> {
-  return new Promise((resolve) => setTimeout(resolve, ms));
 }
 
 function loadArtifact(artifactPath: string): ContractArtifact {
@@ -364,178 +362,6 @@ function getConfig(): SmokeConfig {
   }
 
   return config;
-}
-
-function startManagedProcess(
-  name: string,
-  command: string,
-  args: string[],
-  options: {
-    cwd: string;
-    env?: NodeJS.ProcessEnv;
-  },
-): ManagedProcess {
-  let logs = "";
-  const processHandle = spawn(command, args, {
-    cwd: options.cwd,
-    env: options.env ?? process.env,
-    detached: true,
-    stdio: ["ignore", "pipe", "pipe"],
-  });
-
-  processHandle.stdout.on("data", (chunk: Buffer) => {
-    const text = chunk.toString();
-    logs += text;
-    process.stdout.write(`[${name}] ${text}`);
-  });
-  processHandle.stderr.on("data", (chunk: Buffer) => {
-    const text = chunk.toString();
-    logs += text;
-    process.stderr.write(`[${name}] ${text}`);
-  });
-
-  const managed: ManagedProcess = {
-    name,
-    process: processHandle,
-    getLogs: () => logs,
-  };
-  managedProcessRegistry.add(managed);
-  processHandle.on("exit", () => {
-    managedProcessRegistry.delete(managed);
-  });
-  return managed;
-}
-
-async function stopManagedProcess(proc: ManagedProcess): Promise<void> {
-  if (proc.process.exitCode !== null) {
-    managedProcessRegistry.delete(proc);
-    return;
-  }
-
-  const pid = proc.process.pid;
-  let signaled = false;
-  if (typeof pid === "number" && pid > 0) {
-    try {
-      process.kill(-pid, "SIGTERM");
-      signaled = true;
-    } catch {
-      signaled = false;
-    }
-  }
-  if (!signaled) {
-    try {
-      proc.process.kill("SIGTERM");
-    } catch {
-      managedProcessRegistry.delete(proc);
-      return;
-    }
-  }
-  const deadline = Date.now() + 5_000;
-  while (Date.now() < deadline) {
-    if (proc.process.exitCode !== null) {
-      managedProcessRegistry.delete(proc);
-      return;
-    }
-    await sleep(100);
-  }
-  if (typeof pid === "number" && pid > 0) {
-    try {
-      process.kill(-pid, "SIGKILL");
-      managedProcessRegistry.delete(proc);
-      return;
-    } catch {
-      // Fallback to direct child kill if process groups are unavailable.
-    }
-  }
-  try {
-    proc.process.kill("SIGKILL");
-  } catch {
-    // Process may have already exited between checks.
-  }
-  managedProcessRegistry.delete(proc);
-}
-
-async function stopAllManagedProcesses(): Promise<void> {
-  for (const proc of Array.from(managedProcessRegistry).reverse()) {
-    await stopManagedProcess(proc);
-  }
-}
-
-function installManagedProcessSignalHandlers(): void {
-  const handleSignal = (signal: NodeJS.Signals) => {
-    if (shutdownInProgress) {
-      return;
-    }
-    shutdownInProgress = true;
-    void (async () => {
-      console.error(
-        `[services-smoke] Received ${signal}; stopping managed processes...`,
-      );
-      await stopAllManagedProcesses();
-      process.exit(signal === "SIGINT" ? 130 : 143);
-    })();
-  };
-
-  process.once("SIGINT", () => handleSignal("SIGINT"));
-  process.once("SIGTERM", () => handleSignal("SIGTERM"));
-}
-
-async function waitForNodeReady(
-  node: ReturnType<typeof createAztecNodeClient>,
-  timeoutMs: number,
-): Promise<void> {
-  await Promise.race([
-    waitForNode(node),
-    new Promise((_, reject) =>
-      setTimeout(
-        () =>
-          reject(
-            new Error(`Timed out waiting for Aztec node after ${timeoutMs}ms`),
-          ),
-        timeoutMs,
-      ),
-    ),
-  ]);
-}
-
-async function waitForHealth(url: string, timeoutMs: number): Promise<void> {
-  const deadline = Date.now() + timeoutMs;
-  while (Date.now() <= deadline) {
-    try {
-      const response = await fetch(url);
-      if (response.ok) {
-        return;
-      }
-    } catch {
-      // Keep retrying during boot.
-    }
-    await sleep(500);
-  }
-  throw new Error(`Timed out waiting for health endpoint: ${url}`);
-}
-
-async function waitForLog(
-  proc: ManagedProcess,
-  expected: string,
-  timeoutMs: number,
-): Promise<void> {
-  const deadline = Date.now() + timeoutMs;
-  while (Date.now() <= deadline) {
-    if (proc.getLogs().includes(expected)) {
-      return;
-    }
-    if (proc.process.exitCode !== null) {
-      throw new Error(
-        `Process ${proc.name} exited before logging "${expected}" (exit=${proc.process.exitCode})`,
-      );
-    }
-    await sleep(250);
-  }
-  throw new Error(
-    `Timed out waiting for ${proc.name} log "${expected}". Recent logs:\n${proc
-      .getLogs()
-      .slice(-4000)}`,
-  );
 }
 
 async function waitForTopupBridgeOutcome(
@@ -1778,7 +1604,7 @@ async function runCreditFeeScenario(
 }
 
 async function main() {
-  installManagedProcessSignalHandlers();
+  installManagedProcessSignalHandlers("services-smoke");
   const config = getConfig();
   const scriptDir =
     typeof __dirname === "string"
